# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/guest/getlist/{ballroomId}": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Get guest lists from Emka Backend",
                "description": "Emka backend is a layer from application mother.",
                "operationId": "d91c908c29a67e5a3f20cd467a70e1a0",
                "parameters": [
                    {
                        "name": "ballroomId",
                        "in": "path",
                        "description": "ID of the ballroom to retrieve guests for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ballroom-123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of guests retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "guest_id": {
                                                "type": "string",
                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                            },
                                            "element_id": {
                                                "type": "string",
                                                "example": "chair-123"
                                            },
                                            "ballroom_id": {
                                                "type": "string",
                                                "example": "ballroom-123"
                                            },
                                            "seated": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "group": {
                                                "type": "string",
                                                "example": "VIP"
                                            },
                                            "age_group": {
                                                "type": "string",
                                                "example": "children"
                                            },
                                            "gender": {
                                                "type": "string",
                                                "example": "M"
                                            },
                                            "menu": {
                                                "type": "string",
                                                "example": "S"
                                            },
                                            "parameters": {
                                                "type": "object",
                                                "example": {
                                                    "dietary_preference": "vegetarian"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials/token"
                    },
                    "404": {
                        "description": "Ballroom not found"
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "summary": "Register a new user. You should provide following parameters, describe later in parameters section: name, email, password and c_password(repeated)",
                "operationId": "25dc775588d2e73b509bcb375d425b27",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "c_password",
                        "in": "query",
                        "description": "User's password confirmed",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Authenticate user and generate JWT token. Use this token later for each call to authorize",
                "operationId": "82029854b7191d64582592c46e594dfd",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}